+++ src/Predictor.Models/DataStructures/Hourly/Temperature.cs
@@ -3,7 +3,7 @@ using Microsoft.ML.Data;
 
 namespace Predictor.Models.DataStructures.Hourly
 {
-    public class Temperature : Cities<double>
+    public class Temperature : Cities<float>
     {
         [LoadColumn(0)]
         public DateTime DateTime { get; set; }
